Kaggriculture: move pure view-building fns to utils.ts
